utils refactor and eth_estimate gas tests
