+++ packages/eth-rpc-adapter/package.json
@@ -14,7 +14,7 @@
     "start": "ts-node -r tsconfig-paths/register src/index.ts",
     "start:coverage": "COVERAGE_DIR=${COVERAGE_DIR:-eth-rpc-adapter/coverage} pm2 start --name eth-rpc 'nyc --report-dir=$COVERAGE_DIR node -r ts-node/register -r tsconfig-paths/register src/index.ts' && yarn health-check",
     "stop:coverage": "pm2 stop eth-rpc && pm2 delete eth-rpc",
-    "test": "vitest",
+    "test": "vitest --no-file-parallelism",
     "test:coverage": "COVERAGE_DIR=${COVERAGE_DIR:-eth-rpc-adapter/coverage} yarn start:coverage && COVERAGE_DIR=${COVERAGE_DIR:-eth-rpc-adapter/coverage} yarn test --run && yarn stop:coverage"
   },
   "peerDependencies": {
