+++ packages/eth-rpc-adapter/src/__tests__/utils/consts.ts
@@ -1,4 +1,4 @@
-export const NODE_RPC_URL = process.env.ENDPOINT_URL || 'ws://127.0.0.1:9944';
+export const NODE_URL = process.env.ENDPOINT_URL || 'ws://127.0.0.1:9944';
 export const KARURA_ETH_RPC_URL = process.env.KARURA_ETH_RPC_URL || 'http://127.0.0.1:8546';
 export const ETH_RPC_URL = process.env.ETH_RPC_URL || 'http://127.0.0.1:8545';
 export const ETH_RPC_URL_WS = process.env.ETH_RPC_URL_WS || ETH_RPC_URL.replace('http', 'ws');
@@ -26,223 +26,3 @@ export const KARURA_SEND_KAR_TX_HASH = '0x69493fd597760d5ad3a81ebbbb48abcc686d33
 
 export const deployHelloWorldData =
   '0x60806040526040518060400160405280600c81526020017f48656c6c6f20576f726c642100000000000000000000000000000000000000008152506000908051906020019061004f929190610062565b5034801561005c57600080fd5b50610166565b82805461006e90610134565b90600052602060002090601f01602090048101928261009057600085556100d7565b82601f106100a957805160ff19168380011785556100d7565b828001600101855582156100d7579182015b828111156100d65782518255916020019190600101906100bb565b5b5090506100e491906100e8565b5090565b5b808211156101015760008160009055506001016100e9565b5090565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b6000600282049050600182168061014c57607f821691505b602082108114156101605761015f610105565b5b50919050565b61022e806101756000396000f3fe608060405234801561001057600080fd5b506004361061002b5760003560e01c8063c605f76c14610030575b600080fd5b61003861004e565b6040516100459190610175565b60405180910390f35b6000805461005b906101c6565b80601f0160208091040260200160405190810160405280929190818152602001828054610087906101c6565b80156100d45780601f106100a9576101008083540402835291602001916100d4565b820191906000526020600020905b8154815290600101906020018083116100b757829003601f168201915b505050505081565b600081519050919050565b600082825260208201905092915050565b60005b838110156101165780820151818401526020810190506100fb565b83811115610125576000848401525b50505050565b6000601f19601f8301169050919050565b6000610147826100dc565b61015181856100e7565b93506101618185602086016100f8565b61016a8161012b565b840191505092915050565b6000602082019050818103600083015261018f818461013c565b905092915050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b600060028204905060018216806101de57607f821691505b602082108114156101f2576101f1610197565b5b5091905056fea26469706673582212204d363ed34111d1be492d4fd086e9f2df62b3c625e89ade31f30e63201ed1e24f64736f6c63430008090033';
-
-export const log6 = {
-  blockNumber: '0x6',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000e8d4a51000000000000000000000000000000000000000000000000000000001d131f6171f000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000001',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x7b1ccce9b5299ff0ae3d9adc0855268a4ad3527b2bcde01ccadde2fb878ecb8a',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-  ],
-};
-
-export const log7 = {
-  blockNumber: '0x7',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000e8d4a51000000000000000000000000000000000000000000000000000000000094b686d800000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000010000000000000000000000000000000000000000000100000000000000000002',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x7b1ccce9b5299ff0ae3d9adc0855268a4ad3527b2bcde01ccadde2fb878ecb8a',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-  ],
-};
-
-export const log8 = {
-  blockNumber: '0x8',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x000000000000000000000000000000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000001',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x7b1ccce9b5299ff0ae3d9adc0855268a4ad3527b2bcde01ccadde2fb878ecb8a',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-  ],
-};
-
-export const log9 = {
-  blockNumber: '0x9',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000001a00000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000003000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001000000000000000000010000000000000000000000000000000000000000000100000000000000000002',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x7b1ccce9b5299ff0ae3d9adc0855268a4ad3527b2bcde01ccadde2fb878ecb8a',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-  ],
-};
-
-export const log10 = {
-  blockNumber: '0xa',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x0000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000748849ea0c000000000000000000000000000000000000000000000000000000e8d4a51000000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000001',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x7b1ccce9b5299ff0ae3d9adc0855268a4ad3527b2bcde01ccadde2fb878ecb8a',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-  ],
-};
-
-export const log11 = {
-  blockNumber: '0xb',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x000000000000000000000000000000000000000000000000000000e8d4a51000000000000000000000000000000000000000000000000000000000e8d4a51000',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x5b6f5f6550282279c4e72b95a8ba538bea92c64dec9e8c7c08a556d4457225c8',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-    '0x0000000000000000000000000000000000000000000100000000000000000000',
-    '0x0000000000000000000000000000000000000000000100000000000000000001',
-  ],
-};
-
-export const log12 = {
-  blockNumber: '0xc',
-  transactionIndex: '0x0',
-  address: '0x0000000000000000000000000000000000000803',
-  data: '0x000000000000000000000000000000000000000000000000000000174876e800',
-  logIndex: '0x0',
-  removed: false,
-  topics: [
-    '0x038116623990e7d0fed04a27e35b5dc88000ea942b37360c5898ae750bfa5df6',
-    '0x0000000000000000000000000230135fded668a3f7894966b14f42e65da322e4',
-    '0x0000000000000000000000000000000000000000000100000000000000000000',
-    '0x0000000000000000000000000000000000000000000100000000000000000001',
-  ],
-};
-
-export const log13 = {
-  blockNumber: '0xd',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0xe85ef9063dd28f157eb97ca03f50f4a3bdecd37e',
-  data: '0x0000000000000000000000000000000000000000000000000de0b6b3a7640000',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x0000000000000000000000000000000000000000000000000000000000000000',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-  ],
-  logIndex: '0x0',
-};
-
-export const log14 = {
-  blockNumber: '0xe',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0x532394de2ca885b7e0306a2e258074cca4e42449',
-  data: '0x0000000000000000000000000000000000000000000000000de0b6b3a7640000',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x0000000000000000000000000000000000000000000000000000000000000000',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-  ],
-  logIndex: '0x0',
-};
-
-export const log20_0 = {
-  blockNumber: '0x14',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0x532394de2ca885b7e0306a2e258074cca4e42449',
-  data: '0x0000000000000000000000000000000000000000000000000000000000002710',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-    '0x000000000000000000000000905c015e38c24ed973fd6075541a124c621fa743',
-  ],
-  logIndex: '0x0',
-};
-
-export const log20_1 = {
-  blockNumber: '0x14',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0xe85ef9063dd28f157eb97ca03f50f4a3bdecd37e',
-  data: '0x00000000000000000000000000000000000000000000000000000000000003e8',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-    '0x000000000000000000000000905c015e38c24ed973fd6075541a124c621fa743',
-  ],
-  logIndex: '0x1',
-};
-
-export const log22_0 = {
-  blockNumber: '0x16',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0xe85ef9063dd28f157eb97ca03f50f4a3bdecd37e',
-  data: '0x000000000000000000000000000000000000000000000000000000000000000a',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-    '0x000000000000000000000000905c015e38c24ed973fd6075541a124c621fa743',
-  ],
-  logIndex: '0x0',
-};
-
-export const log22_1 = {
-  blockNumber: '0x16',
-  transactionIndex: '0x0',
-  removed: false,
-  address: '0x532394de2ca885b7e0306a2e258074cca4e42449',
-  data: '0x0000000000000000000000000000000000000000000000000000000000000062',
-  topics: [
-    '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-    '0x000000000000000000000000905c015e38c24ed973fd6075541a124c621fa743',
-    '0x00000000000000000000000082a258cb20e2adb4788153cd5eb5839615ece9a0',
-  ],
-  logIndex: '0x1',
-};
-
-export const allLogs = [
-  log6,
-  log7,
-  log8,
-  log9,
-  log10,
-  log11,
-  log12,
-  log13,
-  log14,
-  log20_0,
-  log20_1,
-  log22_0,
-  log22_1,
-] as LogHexified[];
-
-export const ERC20_BYTECODE = '6060604052341561000f57600080fd5b604051610dd1380380610dd18339810160405280805190602001909190805182019190602001805190602001909190805182019190505083600160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508360008190555082600390805190602001906100a79291906100e3565b5081600460006101000a81548160ff021916908360ff16021790555080600590805190602001906100d99291906100e3565b5050505050610188565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061012457805160ff1916838001178555610152565b82800160010185558215610152579182015b82811115610151578251825591602001919060010190610136565b5b50905061015f9190610163565b5090565b61018591905b80821115610181576000816000905550600101610169565b5090565b90565b610c3a806101976000396000f3006060604052600436106100af576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff16806306fdde03146100b4578063095ea7b31461014257806318160ddd1461019c57806323b872dd146101c557806327e235e31461023e578063313ce5671461028b5780635c658165146102ba57806370a082311461032657806395d89b4114610373578063a9059cbb14610401578063dd62ed3e1461045b575b600080fd5b34156100bf57600080fd5b6100c76104c7565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156101075780820151818401526020810190506100ec565b50505050905090810190601f1680156101345780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b341561014d57600080fd5b610182600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091908035906020019091905050610565565b604051808215151515815260200191505060405180910390f35b34156101a757600080fd5b6101af610657565b6040518082815260200191505060405180910390f35b34156101d057600080fd5b610224600480803573ffffffffffffffffffffffffffffffffffffffff1690602001909190803573ffffffffffffffffffffffffffffffffffffffff1690602001909190803590602001909190505061065d565b604051808215151515815260200191505060405180910390f35b341561024957600080fd5b610275600480803573ffffffffffffffffffffffffffffffffffffffff169060200190919050506108f7565b6040518082815260200191505060405180910390f35b341561029657600080fd5b61029e61090f565b604051808260ff1660ff16815260200191505060405180910390f35b34156102c557600080fd5b610310600480803573ffffffffffffffffffffffffffffffffffffffff1690602001909190803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610922565b6040518082815260200191505060405180910390f35b341561033157600080fd5b61035d600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610947565b6040518082815260200191505060405180910390f35b341561037e57600080fd5b610386610990565b6040518080602001828103825283818151815260200191508051906020019080838360005b838110156103c65780820151818401526020810190506103ab565b50505050905090810190601f1680156103f35780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b341561040c57600080fd5b610441600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091908035906020019091905050610a2e565b604051808215151515815260200191505060405180910390f35b341561046657600080fd5b6104b1600480803573ffffffffffffffffffffffffffffffffffffffff1690602001909190803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610b87565b6040518082815260200191505060405180910390f35b60038054600181600116156101000203166002900480601f01602080910402602001604051908101604052809291908181526020018280546001816001161561010002031660029004801561055d5780601f106105325761010080835404028352916020019161055d565b820191906000526020600020905b81548152906001019060200180831161054057829003601f168201915b505050505081565b600081600260003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508273ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925846040518082815260200191505060405180910390a36001905092915050565b60005481565b600080600260008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905082600160008773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020541015801561072e5750828110155b151561073957600080fd5b82600160008673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000206000828254019250508190555082600160008773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825403925050819055507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff8110156108865782600260008773ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825403925050819055505b8373ffffffffffffffffffffffffffffffffffffffff168573ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef856040518082815260200191505060405180910390a360019150509392505050565b60016020528060005260406000206000915090505481565b600460009054906101000a900460ff1681565b6002602052816000526040600020602052806000526040600020600091509150505481565b6000600160008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b60058054600181600116156101000203166002900480601f016020809104026020016040519081016040528092919081815260200182805460018160011615610100020316600290048015610a265780601f106109fb57610100808354040283529160200191610a26565b820191906000526020600020905b815481529060010190602001808311610a0957829003601f168201915b505050505081565b600081600160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000205410151515610a7e57600080fd5b81600160003373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000206000828254039250508190555081600160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508273ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef846040518082815260200191505060405180910390a36001905092915050565b6000600260008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050929150505600a165627a7a72305820df254047bc8f2904ad3e966b6db116d703bebd40efadadb5e738c836ffc8f58a0029';
-
-/* -----
-   contract GasMonster {
-       uint256 x = 1;
-       function run() external {
-           require(gasleft() > 1010000, "gaslimit not enough");
-           x = x + 1;
-       }
-   }
-                                                          ----- */
-export const GASMONSTER_BYTECODE = '6080604052600160005534801561001557600080fd5b506101d8806100256000396000f3fe608060405234801561001057600080fd5b506004361061002b5760003560e01c8063c040622614610030575b600080fd5b61003861003a565b005b620f69505a1161007f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610076906100b9565b60405180910390fd5b600160005461008e91906100ea565b600081905550565b60006100a36013836100d9565b91506100ae82610179565b602082019050919050565b600060208201905081810360008301526100d281610096565b9050919050565b600082825260208201905092915050565b60006100f582610140565b915061010083610140565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff038211156101355761013461014a565b5b828201905092915050565b6000819050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f6761736c696d6974206e6f7420656e6f7567680000000000000000000000000060008201525056fea26469706673582212203f50c8d7ac2c3bcf6ebf8adbadba541fff53382a169109676559aabee8a78ddf64736f6c63430008070033';
-
-export const GASMONSTER_ABI = [
-  {
-    inputs: [],
-    name: 'run',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-];
-
-export const GAS_MONSTER_GAS_REQUIRED = 1010000;
