+++ packages/eth-rpc-adapter/src/__tests__/utils/gas-monster.ts
@@ -1,0 +1,32 @@
+import { ContractFactory, Signer } from 'ethers';
+
+/* -----
+   contract GasMonster {
+       uint256 x = 1;
+       function run() external {
+           require(gasleft() > 1010000, "gaslimit not enough");
+           x = x + 1;
+       }
+   }
+                                                          ----- */
+export const GASMONSTER_BYTECODE = '6080604052600160005534801561001557600080fd5b506101d8806100256000396000f3fe608060405234801561001057600080fd5b506004361061002b5760003560e01c8063c040622614610030575b600080fd5b61003861003a565b005b620f69505a1161007f576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610076906100b9565b60405180910390fd5b600160005461008e91906100ea565b600081905550565b60006100a36013836100d9565b91506100ae82610179565b602082019050919050565b600060208201905081810360008301526100d281610096565b9050919050565b600082825260208201905092915050565b60006100f582610140565b915061010083610140565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff038211156101355761013461014a565b5b828201905092915050565b6000819050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f6761736c696d6974206e6f7420656e6f7567680000000000000000000000000060008201525056fea26469706673582212203f50c8d7ac2c3bcf6ebf8adbadba541fff53382a169109676559aabee8a78ddf64736f6c63430008070033';
+
+export const GASMONSTER_ABI = [
+  {
+    inputs: [],
+    name: 'run',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+];
+
+export const GAS_MONSTER_GAS_REQUIRED = 1010000;
+
+export const deployGasMonster = async (wallet: Signer) => {
+  const GM = new ContractFactory(GASMONSTER_ABI, GASMONSTER_BYTECODE, wallet);
+  const gm = await GM.deploy();
+  await gm.deployed();
+
+  return gm;
+};
