+++ packages/eth-rpc-adapter/src/__tests__/utils/index.ts
@@ -1,4 +1,7 @@
-export * from './SubsManager';
+export * from './subs-manager';
 export * from './test-utils';
-export * from './evmAccounts';
+export * from './evm-accounts';
 export * from './consts';
+export * from './gas-monster';
+export * from './eth-rpc-apis';
+export * from './erc20';
