+++ packages/eth-rpc-adapter/vitest.config.ts
@@ -5,11 +5,11 @@ export default mergeConfig(
   configShared,
   {
     test: {
-      // include: ['**/*.test.ts'],  // TODO: run all tests
-      include: [
-        'src/__tests__/signer.test.ts',
-        'src/__tests__/subscription.test.ts',
-      ],
+      include: ['**/*.test.ts'],  // TODO: run all tests
+      // include: [
+      //   'src/__tests__/signer.test.ts',
+      //   'src/__tests__/subscription.test.ts',
+      // ],
     },
   }
 );
